Pixel access returns double instead of float
git-svn-id: file:///home/roberto/Git/svn@9681 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 2ec6d34fb2886cd81d20e371df259a057dbaf0cb

+++ trunk/xmipp/libraries/data/micrograph.h
@@ -381,7 +381,7 @@ public:
     /** Pixel access for reading.
         These coordinates follow the physical Xmipp convention
         {../../../Extra_Docs/Conventions.html} for coordinates */
-    float operator()(int y, int x) const
+    double operator()(int y, int x) const
     {
         if (y < 0 || y >= Ydim || x < 0 || x >= Xdim)
             // COSS: REPORT_ERROR(1, "Micrograph::(): index out of range");
